requirements.txt
asn1crypto==1.5.1
asttokens==3.0.1
boto3==1.42.60
botocore==1.42.60
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
cryptography==46.0.5
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
filelock==3.25.0
idna==3.11
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.1.0
jupyter_client==8.8.0
jupyter_core==5.9.1
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.4.3
packaging==26.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
platformdirs==4.9.2
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==3.0
Pygments==2.19.2
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
pyzmq==27.1.0
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
snowflake-connector-python==4.3.0
sortedcontainers==2.4.0
stack-data==0.6.3
tomlkit==0.14.0
tornado==6.5.5
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
